% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_correction.R
\name{bias_correction}
\alias{bias_correction}
\title{Bias Correction for Simulated Climate Data}
\usage{
bias_correction(
  simulated_data,
  observed_data,
  variables = c("Tmin", "Tmax", "Avg.Temp", "Rainfall", "RH", "WindSpeed",
    "Solar_Radiation", "ET0"),
  method = "mean_scaling",
  digits = 2,
  return_factors = TRUE
)
}
\arguments{
\item{simulated_data}{Simulated climate data frame.}

\item{observed_data}{Observed/reference climate data frame.}

\item{variables}{Character vector of variables to correct.}

\item{method}{Bias-correction method.
Options are:
\code{"mean_scaling"},
\code{"additive"},
\code{"multiplicative"}.}

\item{digits}{Number of decimal places.}

\item{return_factors}{Logical. If TRUE, returns correction
factors and bias statistics.}
}
\value{
A list containing:

\describe{

\item{corrected_data}{
Bias-corrected climate dataset.
}

\item{correction_factors}{
Bias-correction factors and bias statistics applied to
each variable.
}

}
}
\description{
Applies statistical bias correction to simulated climate
variables using observed reference data.
}
\details{
Supported correction methods include:
\itemize{
\item Mean scaling
\item Additive correction
\item Multiplicative correction
}

Physical constraints are automatically enforced after
correction to ensure climatological realism.
}
\examples{
\dontrun{

bc <- bias_correction(
  simulated_data = cd,
  observed_data = obs_data,
  variables = c("Rainfall", "Avg.Temp")
)

head(bc$corrected_data)
bc$correction_factors

}

}
