% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_imports.R
\docType{package}
\name{CDSimX-package}
\alias{CDSimX}
\alias{CDSimX-package}
\title{CDSimX: Simulating Climate Data for Research and Modelling}
\description{
Advanced climate simulation, forecasting, visualization, export, and machine learning tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ikemillar/CDSimX}
  \item Report bugs at \url{https://github.com/ikemillar/CDSimX/issues}
}

}
\author{
\strong{Maintainer}: Isaac Osei \email{ikemillar65@gmail.com}

Authors:
\itemize{
  \item Acheampong Baafi-Adomako \email{acheampongbaafi1@gmail.com}
  \item Jayanti Maji
}

}
\keyword{internal}
