% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_dependence.R
\name{copula_dependence}
\alias{copula_dependence}
\title{Apply Copula Dependence Structure to Climate Variables}
\usage{
copula_dependence(
  climate_data,
  variables = c("Tmin", "Tmax", "Avg.Temp", "Rainfall", "RH", "WindSpeed",
    "Solar_Radiation", "ET0"),
  copula_type = "gaussian",
  df = 4,
  seed = 123,
  digits = 2
)
}
\arguments{
\item{climate_data}{Climate data frame.}

\item{variables}{Variables used in dependence modeling.}

\item{copula_type}{Type of copula.
Options are:
\code{"gaussian"} or \code{"t"}.}

\item{df}{Degrees of freedom for t copula.
Default is 4.}

\item{seed}{Random seed for reproducibility.}

\item{digits}{Number of decimal places.}
}
\value{
A list containing:

\describe{

\item{adjusted_data}{
Climate dataset with copula-adjusted dependence structure.
}

\item{correlation_matrix}{
Empirical correlation matrix used in copula fitting.
}

\item{copula_type}{
Copula family used.
}

}
}
\description{
Introduces multivariate dependence among simulated climate
variables using Gaussian or t copulas.
}
\details{
This function improves realism by preserving correlations
and inter-variable dependency structures commonly observed
in climate systems.

Supported variables include:
\itemize{
\item Tmin
\item Tmax
\item Avg.Temp
\item Rainfall
\item RH
\item WindSpeed
\item Solar_Radiation
\item ET0
}
}
\examples{
\dontrun{

cp <- copula_dependence(cd)

head(cp$adjusted_data)

}

}
