% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_csv.R
\name{export_csv}
\alias{export_csv}
\title{Export CDSimX Data to CSV}
\usage{
export_csv(data, file = "CDSimX_output.csv", row_names = FALSE)
}
\arguments{
\item{data}{A dataframe to export.}

\item{file}{Output CSV filename.}

\item{row_names}{Logical. Include row names?}
}
\value{
Invisibly returns the file path.
}
\description{
Exports any CDSimX dataframe to a CSV file.
}
\examples{
\dontrun{
export_csv(climate_data)
}
}
