% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_climate.R
\name{simulate_climate}
\alias{simulate_climate}
\title{Simulate Integrated Climate Dataset}
\usage{
simulate_climate(stations, time_index, seed = NULL)
}
\arguments{
\item{stations}{Output from
create_stations().}

\item{time_index}{Output from
generate_time_index().}

\item{seed}{Optional random seed
for reproducibility.}
}
\value{
A merged dataframe containing
all simulated climate variables.
}
\description{
Generates a complete synthetic climate dataset
by internally calling all climate simulation
modules and merging their outputs into a
single dataframe.
}
\details{
The function currently simulates:
\itemize{
\item temperature
\item rainfall
\item relative humidity
\item dew point
\item wind speed
\item wind direction
\item solar radiation
\item evapotranspiration
}
}
\examples{
\dontrun{

stations <- create_stations(
  n = 10
)

tindex <- generate_time_index(
  start_date = "2020-01-01",
  end_date   = "2020-12-31",
  frequency  = "monthly"
)

climate <- simulate_climate(
  stations,
  tindex
)

head(climate)

}

}
